feat: add api for products

+++ myProject/src/components/authorization/DeleteUserBtn.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router';
 import { auth } from '../../config/firebase';
 import { clearUser } from '../../store/AuthSlice';
-import { BtnAuth } from '../../styles/Buttons.styled';
+import { BtnDelete } from '../../styles/Buttons.styled';
 
 
 export default function DeleteUserBtn() {
@@ -12,24 +12,28 @@ export default function DeleteUserBtn() {
   const navigate = useNavigate();
 
   function deleteUser() {
-    const user = auth.currentUser;
-    if (user) {
-      user.delete()
-        .then(() => {
-          dispatch(clearUser());
-          navigate('/register');
-        })
-        .catch((error) => {
-          console.error('Error for deleting user', error);
-        });
-    } else {
-      console.error('No have authorized user');
+    const agree = prompt('delete user?')
+    if (agree) {
+      const user = auth.currentUser;
+      if (user) {
+        user.delete()
+          .then(() => {
+            dispatch(clearUser());
+            navigate('/register');
+          })
+          .catch((error) => {
+            console.error('Error for deleting user', error);
+          });
+      } else {
+        console.error('No have authorized user');
+      }
     }
+   
   };
 
   return (
-    <BtnAuth onClick={deleteUser}>
+    <BtnDelete onClick={deleteUser}>
       Delete Account
-    </BtnAuth>
+    </BtnDelete>
   );
 }
